backend/requirements.txt
fastapi
uvicorn[standard]
motor==3.1.1
pymongo==4.3.3
google-generativeai==0.3.2
chromadb==0.4.18
langdetect==1.0.9
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
numpy==1.26.4

